feat(forum_upvote): renaming www app
