+++ config/settings/base.py
@@ -73,6 +73,7 @@ LOCAL_APPS = [
     "lacommunaute.www.forum_views",
     "lacommunaute.www.forum_conversation_views",
     "lacommunaute.www.forum_member_views",
+    "lacommunaute.www.forum_upvote_views",
 ]
 
 INSTALLED_APPS = DJANGO_APPS + LOCAL_APPS + THIRD_PARTIES_APPS
