+++ lacommunaute/www/forum_upvote_views/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from lacommunaute.www.forum_upvote.views import PostUpvoteView
+from lacommunaute.www.forum_upvote_views.views import PostUpvoteView
 
 
 app_name = "forum_upvote"
